following STM32CubeIDE conventions
