+++ .project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>STM32-FreeRTOS-TIM</name>
+	<name>STM32-FreeRTOS-Framework</name>
 	<comment></comment>
 	<projects>
 	</projects>
