+++ .settings/stm32cubeide.project.prefs
@@ -1,5 +1,5 @@
 635E684B79701B039C64EA45C3F84D30=2BD7D4602A0AB9B75FB38728DABF9704
-66BE74F758C12D739921AEA421D593D3=1
+66BE74F758C12D739921AEA421D593D3=3
 8DF89ED150041C4CBC7CB9A9CAA90856=BDB507BA32D9E9EDD44CDC05C27AC7F4
 DC22A860405A8BF2F2C095E5B6529F12=BDB507BA32D9E9EDD44CDC05C27AC7F4
 eclipse.preferences.version=1
