--- Core/Inc/ErrorHandler.h
@@ -1,13 +1,0 @@
-/*
- * ErrorHandler.h
- *
- *  Created on: May 31, 2021
- *      Author: frank
- */
-
-#ifndef INC_ERROR_HANDLER_H_
-#define INC_ERROR_HANDLER_H_
-
-void Error_Handler(void);
-
-#endif
