--- Core/Inc/Init/Init_MxEthernet.h
@@ -1,13 +1,0 @@
-/*
- * Init_MxEthernet.h
- *
- *  Created on: May 31, 2021
- *      Author: frank
- */
-
-#ifndef INC_INIT_INIT_MXETHERNET_H_
-#define INC_INIT_INIT_MXETHERNET_H_
-
-void MX_ETH_Init(void);
-
-#endif /* INC_INIT_INIT_MXETHERNET_H_ */
