--- Core/Inc/Init/Init_MxGpio.h
@@ -1,13 +1,0 @@
-/*
- * Init_MxGpio.h
- *
- *  Created on: May 31, 2021
- *      Author: frank
- */
-
-#ifndef INC_INIT_INIT_MXGPIO_H_
-#define INC_INIT_INIT_MXGPIO_H_
-
-void MX_GPIO_Init(void);
-
-#endif /* INC_INIT_INIT_MXGPIO_H_ */
