--- Core/Inc/Init/Init_MxUsbOtgFsPcdInit.h
@@ -1,13 +1,0 @@
-/*
- * Init_MxUsbOtgFsPcdInit.h
- *
- *  Created on: May 31, 2021
- *      Author: frank
- */
-
-#ifndef INC_INIT_INIT_MXUSBOTGFSPCDINIT_H_
-#define INC_INIT_INIT_MXUSBOTGFSPCDINIT_H_
-
-void MX_USB_OTG_FS_PCD_Init(void);
-
-#endif /* INC_INIT_INIT_MXUSBOTGFSPCDINIT_H_ */
