--- Core/Inc/Init/Init_SystemClock.h
@@ -1,13 +1,0 @@
-/*
- * SystemClock.h
- *
- *  Created on: May 31, 2021
- *      Author: frank
- */
-
-#ifndef SRC_INIT_SYSTEMCLOCK_H_
-#define SRC_INIT_SYSTEMCLOCK_H_
-
-void SystemClock_Config(void);
-
-#endif /* SRC_INIT_SYSTEMCLOCK_H_ */
