--- Core/Inc/Init/Init_Usart3Uart.h
@@ -1,13 +1,0 @@
-/*
- * Init_Usart3Uart.h
- *
- *  Created on: May 31, 2021
- *      Author: frank
- */
-
-#ifndef INC_INIT_INIT_USART3UART_H_
-#define INC_INIT_INIT_USART3UART_H_
-
-void MX_USART3_UART_Init(void);
-
-#endif /* INC_INIT_INIT_USART3UART_H_ */
