--- Core/Inc/Init/Tasks/Init_Task_Default.h
@@ -1,13 +1,0 @@
-/*
- * Init_Task_Led1.h
- *
- *  Created on: Jun 3, 2021
- *      Author: frank
- */
-
-#ifndef INC_INIT_INIT_TASK_DEFDAULT_H_
-#define INC_INIT_INIT_TASK_DEFDAULT_H_
-
-void Init_Task_Default();
-
-#endif /* INC_INIT_INIT_TASK_DEFDAULT_H_ */
