--- Core/Inc/Init/Tasks/Init_Task_Led1.h
@@ -1,13 +1,0 @@
-/*
- * Init_Task_Led1.h
- *
- *  Created on: Jun 3, 2021
- *      Author: frank
- */
-
-#ifndef INC_INIT_INIT_TASK_LED1_H_
-#define INC_INIT_INIT_TASK_LED1_H_
-
-void Init_Task_Led1();
-
-#endif /* INC_INIT_INIT_TASK_LED1_H_ */
