--- Core/Inc/Init/Tasks/Init_Task_Led3.h
@@ -1,13 +1,0 @@
-/*
- * Init_Task_Led1.h
- *
- *  Created on: Jun 3, 2021
- *      Author: frank
- */
-
-#ifndef INC_INIT_INIT_TASK_LED3_H_
-#define INC_INIT_INIT_TASK_LED3_H_
-
-void Init_Task_Led3();
-
-#endif /* INC_INIT_INIT_TASK_LED3_H_ */
