--- Core/Inc/Run/Run_Task_Default.h
@@ -1,13 +1,0 @@
-/*
- * Task_Default.h
- *
- *  Created on: May 31, 2021
- *      Author: frank
- */
-
-#ifndef INC_RUN_RUN_TASK_DEFAULT_H_
-#define INC_RUN_RUN_TASK_DEFAULT_H_
-
-void Run_Task_Default(void const * argument);
-
-#endif /* INC_RUN_RUN_TASK_DEFAULT_H_ */
