--- Core/Inc/Run/Run_Task_Led1.h
@@ -1,13 +1,0 @@
-/*
- * Task_Led1.h
- *
- *  Created on: Jun 1, 2021
- *      Author: frank
- */
-
-#ifndef INC_RUN_RUN_TASK_LED1_H_
-#define INC_RUN_RUN_TASK_LED1_H_
-
-void Run_Task_Led1(void const * argument);
-
-#endif /* INC_RUN_RUN_TASK_LED1_H_ */
