--- Core/Inc/Run/Run_Task_Led3.h
@@ -1,13 +1,0 @@
-/*
- * Task_Led3.h
- *
- *  Created on: Jun 1, 2021
- *      Author: frank
- */
-
-#ifndef INC_RUN_RUN_TASK_LED3_H_
-#define INC_RUN_RUN_TASK_LED3_H_
-
-void Run_Task_Led3(void const * argument);
-
-#endif /* INC_RUN_RUN_TASK_LED3_H_ */
