--- Core/Src/Init/Tasks/Init_Task_Default.c
@@ -1,16 +1,0 @@
-/*
- * Init_Task_Led1.c
- *
- *  Created on: Jun 3, 2021
- *      Author: frank
- */
-
-#include "Init/Init_Global.h"
-
-#include "Init/Tasks/Init_Task_Default.h"
-
-void Init_Task_Default()
-{
-
-}
-
