--- Core/Src/logging.c
@@ -1,8 +1,0 @@
-/*
- * logging.c
- *
- *  Created on: May 31, 2021
- *      Author: frank
- */
-
-
