--- Debug/.gitignore
@@ -1,7 +1,0 @@
-/Core/
-/Drivers/
-/Middlewares/
-/makefile
-/objects.list
-/objects.mk
-/sources.mk
