+++ STM32-FreeRTOS-Framework.ioc
@@ -20,7 +20,7 @@ RCC.PLLQCLKFreq_Value=48000000
 PC5.Locked=true
 RCC.SYSCLKSourceVirtual=RCC_SYSCLKSOURCE_PLLCLK
 PA12.GPIOParameters=GPIO_Label
-ProjectManager.functionlistsort=1-MX_GPIO_Init-GPIO-false-HAL-true,2-SystemClock_Config-RCC-false-HAL-false,3-MX_ETH_Init-ETH-false-HAL-true,4-MX_USART3_UART_Init-USART3-false-HAL-true,5-MX_USB_OTG_FS_PCD_Init-USB_OTG_FS-false-HAL-true,0-MX_CORTEX_M7_Init-CORTEX_M7-false-HAL-true
+ProjectManager.functionlistsort=1-MX_GPIO_Init-GPIO-false-HAL-true,2-SystemClock_Config-RCC-false-HAL-false,3-MX_ETH_Init-ETH-false-HAL-true,3-MX_USART3_UART_Init-USART3-false-HAL-true,4-MX_USB_OTG_FS_PCD_Init-USB_OTG_FS-false-HAL-true,0-MX_CORTEX_M7_Init-CORTEX_M7-false-HAL-true
 PA9.GPIOParameters=GPIO_Label
 PD8.Locked=true
 PA11.GPIOParameters=GPIO_Label
@@ -30,14 +30,13 @@ PB0.GPIO_Label=LD1 [Green]
 PC14/OSC32_IN.Mode=LSE-External-Oscillator
 RCC.PLLI2SRCLKFreq_Value=192000000
 RCC.USART2Freq_Value=48000000
-PC5.Mode=RMII
 PC13.GPIO_Label=USER_Btn [B1]
 PC1.Locked=true
 PB13.GPIO_Label=RMII_TXD1 [LAN8742A-CZ-TR_TXD1]
 PB13.Signal=ETH_TXD1
 PG13.Signal=ETH_TXD0
-PD9.Mode=Asynchronous
 PA2.GPIOParameters=GPIO_Label
+PD9.Mode=Asynchronous
 PinOutPanel.RotationAngle=0
 PG11.Signal=ETH_TX_EN
 RCC.MCO1PinFreq_Value=16000000
@@ -47,22 +46,21 @@ SH.GPXTI13.0=GPIO_EXTI13
 USART3.VirtualMode-Asynchronous=VM_ASYNC
 RCC.I2C3Freq_Value=48000000
 RCC.LPTIM1Freq_Value=48000000
-Mcu.IP4=RCC
+Mcu.IP4=SYS
 RCC.FCLKCortexFreq_Value=96000000
-Mcu.IP5=SYS
-Mcu.IP2=FREERTOS
+Mcu.IP5=TIM14
+Mcu.IP2=NVIC
 NVIC.SVCall_IRQn=true\:0\:0\:false\:false\:false\:false\:true\:false
-Mcu.IP3=NVIC
+Mcu.IP3=RCC
 Mcu.IP0=CORTEX_M7
 PA12.Locked=true
-Mcu.IP1=ETH
+Mcu.IP1=FREERTOS
 PA12.Signal=USB_OTG_FS_DP
 Mcu.UserConstants=
 RCC.LCDTFTFreq_Value=96000000
 PC1.GPIOParameters=GPIO_Label
 RCC.SDMMCFreq_Value=96000000
 SH.GPXTI13.ConfNb=1
-ETH.PHY_Value=0
 Mcu.ThirdPartyNb=0
 PH0/OSC_IN.Mode=HSE-External-Clock-Source
 RCC.HCLKFreq_Value=96000000
@@ -101,8 +99,10 @@ RCC.HSE_VALUE=8000000
 PG13.Locked=true
 NVIC.UsageFault_IRQn=true\:0\:0\:false\:false\:true\:false\:true\:false
 NVIC.DebugMonitor_IRQn=true\:0\:0\:false\:false\:true\:false\:true\:false
+NVIC.TIM8_TRG_COM_TIM14_IRQn=true\:5\:0\:false\:false\:true\:true\:true\:true
 PA8.Locked=true
-NVIC.SysTick_IRQn=true\:15\:0\:false\:false\:true\:true\:true\:true
+TIM14.Period=2000 - 1
+NVIC.SysTick_IRQn=true\:15\:0\:false\:false\:false\:true\:true\:true
 RCC.PLLMUL=RCC_PLL_MUL9
 PH0/OSC_IN.Locked=true
 ProjectManager.FirmwarePackage=STM32Cube FW_F7 V1.16.1
@@ -113,21 +113,20 @@ ProjectManager.BackupPrevious=false
 RCC.VCOInputFreq_Value=2000000
 PC4.GPIO_Label=RMII_RXD0 [LAN8742A-CZ-TR_RXD0]
 PA14.Mode=Serial_Wire
+VP_SYS_VS_tim1.Mode=TIM1
 PH1/OSC_OUT.Locked=true
 File.Version=6
 PA10.GPIO_Label=USB_ID
 PB3.GPIOParameters=GPIO_Label
 PB7.Signal=GPIO_Output
 PA8.Signal=USB_OTG_FS_SOF
-PG13.Mode=RMII
-ETH.PHY_Name=LAN8742A_PHY_ADDRESS
 NVIC.PendSV_IRQn=true\:15\:0\:false\:false\:false\:true\:true\:false
 PC15/OSC32_OUT.Locked=true
 VP_FREERTOS_VS_CMSIS_V1.Mode=CMSIS_V1
 PH1/OSC_OUT.Mode=HSE-External-Clock-Source
 ProjectManager.HalAssertFull=false
 PB0.Locked=true
-ProjectManager.ProjectName=STM32-FreeRTOS-TIM
+ProjectManager.ProjectName=STM32-FreeRTOS-Framework
 RCC.TIM1Freq_Value=72000000
 PA13.GPIOParameters=GPIO_Label
 RCC.PLLMCOFreq_Value=72000000
@@ -136,6 +135,7 @@ RCC.MCO2PinFreq_Value=96000000
 Mcu.Package=LQFP144
 RCC.TIM15Freq_Value=72000000
 PA7.Locked=true
+NVIC.TimeBase=TIM1_UP_TIM10_IRQn
 PG11.Locked=true
 PG13.GPIOParameters=GPIO_Label
 RCC.ADC34outputFreq_Value=72000000
@@ -144,7 +144,7 @@ PA14.Locked=true
 ProjectManager.ToolChainLocation=
 PA2.GPIO_Label=RMII_MDIO [LAN8742A-CZ-TR_MDIO]
 RCC.LSI_VALUE=32000
-VP_SYS_VS_Systick.Signal=SYS_VS_Systick
+NVIC.TimeBaseIP=TIM1
 USART3.IPParameters=VirtualMode-Asynchronous
 PA10.Signal=USB_OTG_FS_ID
 USB_OTG_FS.VirtualMode=Device_Only
@@ -158,26 +158,26 @@ PC4.Locked=true
 PG11.GPIO_Label=RMII_TX_EN [LAN8742A-CZ-TR_TXEN]
 PC5.Signal=ETH_RXD1
 USB_OTG_FS.IPParameters=VirtualMode
-PB13.Mode=RMII
+VP_TIM14_VS_ClockSourceINT.Mode=Enable_Timer
 RCC.DFSDMAudioFreq_Value=192000000
+TIM14.IPParameters=Prescaler,Period
 ProjectManager.CustomerFirmwarePackage=
 PC4.GPIOParameters=GPIO_Label
-PC4.Mode=RMII
 RCC.PLLSAIQCLKFreq_Value=192000000
 PB3.Locked=true
 PA2.Locked=true
 PB3.GPIO_Label=SWO
 RCC.I2SFreq_Value=192000000
 RCC.PLLQoutputFreq_Value=48000000
-ProjectManager.ProjectFileName=STM32-FreeRTOS-TIM.ioc
+ProjectManager.ProjectFileName=STM32-FreeRTOS-Framework.ioc
 RCC.ADC12outputFreq_Value=72000000
 FREERTOS.Tasks01=defaultTask,0,128,StartDefaultTask,Default,NULL,Dynamic,NULL,NULL
 PG7.Locked=true
-PA7.Mode=RMII
 PG7.Signal=GPIO_Input
 RCC.UART7Freq_Value=48000000
 PD9.GPIO_Label=STLK_TX [STM32F103CBT6_PA2]
-Mcu.PinsNb=31
+Mcu.PinsNb=32
+VP_TIM14_VS_ClockSourceINT.Signal=TIM14_VS_ClockSourceINT
 ProjectManager.NoMain=false
 PC13.Locked=true
 NVIC.SavedSvcallIrqHandlerGenerated=true
@@ -195,11 +195,9 @@ RCC.APB1CLKDivider=RCC_HCLK_DIV2
 RCC.LCDTFToutputFreq_Value=96000000
 boardIOC=true
 PG13.GPIO_Label=RMII_TXD0 [LAN8742A-CZ-TR_TXD0]
-ETH.PhyAddress=0
 PA12.GPIO_Label=USB_DP
 RCC.FamilyName=M
 PD8.GPIO_Label=STLK_RX [STM32F103CBT6_PA3]
-ETH.IPParameters=MediaInterface,PHY_Name,PHY_Value,PhyAddress
 RCC.WatchDogFreq_Value=32000
 PA13.Signal=SYS_JTMS-SWDIO
 VP_FREERTOS_VS_CMSIS_V1.Signal=FREERTOS_VS_CMSIS_V1
@@ -228,21 +226,21 @@ PH0/OSC_IN.Signal=RCC_OSC_IN
 NVIC.SavedSystickIrqHandlerGenerated=true
 RCC.APB2Freq_Value=96000000
 RCC.TIM8Freq_Value=72000000
-PA1.Mode=RMII
 RCC.UART4Freq_Value=48000000
 MxCube.Version=6.2.1
 RCC.PRESCALERUSB=RCC_USBCLKSOURCE_PLL_DIV1_5
 RCC.I2C1Freq_Value=48000000
+VP_SYS_VS_tim1.Signal=SYS_VS_tim1
 RCC.PLLI2SPCLKFreq_Value=192000000
 PC13.GPIOParameters=GPIO_Label
 PG7.GPIO_Label=USB_OverCurrent [STMPS2151STR_FAULT]
 RCC.RNGFreq_Value=48000000
-Mcu.Pin30=VP_SYS_VS_Systick
+Mcu.Pin30=VP_SYS_VS_tim1
 PA1.GPIO_Label=RMII_REF_CLK [LAN8742A-CZ-TR_REFCLK0]
 PC1.GPIO_Label=RMII_MDC [LAN8742A-CZ-TR_MDC]
 PA10.GPIOParameters=GPIO_Label
-VP_SYS_VS_Systick.Mode=SysTick
 RCC.EthernetFreq_Value=96000000
+Mcu.Pin31=VP_TIM14_VS_ClockSourceINT
 PA9.Locked=true
 NVIC.NonMaskableInt_IRQn=true\:0\:0\:false\:false\:true\:false\:true\:false
 RCC.UART5Freq_Value=48000000
@@ -256,6 +254,7 @@ Mcu.Pin26=PG13
 Mcu.Pin27=PB3
 NVIC.SavedPendsvIrqHandlerGenerated=true
 RCC.RTCHSEDivFreq_Value=4000000
+TIM14.Prescaler=9600 - 1
 PA2.Signal=ETH_MDIO
 PA13.GPIO_Label=TMS
 Mcu.Pin24=PA14
@@ -270,7 +269,6 @@ RCC.48MHZClocksFreq_Value=24000000
 PB3.Signal=SYS_JTDO-SWO
 RCC.SYSCLKFreq_VALUE=96000000
 Mcu.Pin22=PA12
-ETH.MediaInterface=ETH_MEDIA_INTERFACE_RMII
 Mcu.Pin23=PA13
 PA1.Locked=true
 PA7.GPIO_Label=RMII_CRS_DV [LAN8742A-CZ-TR_CRS_DV]
@@ -287,7 +285,6 @@ NVIC.MemoryManagement_IRQn=true\:0\:0\:false\:false\:true\:false\:true\:false
 PA13.Locked=true
 PC15/OSC32_OUT.Mode=LSE-External-Oscillator
 ProjectManager.CompilerOptimize=6
-PG11.Mode=RMII
 PA11.Signal=USB_OTG_FS_DM
 PA14.Signal=SYS_JTCK-SWCLK
 ProjectManager.HeapSize=0x200
@@ -297,6 +294,7 @@ Mcu.Pin16=PG6
 Mcu.Pin13=PB14
 Mcu.Pin14=PD8
 Mcu.Pin19=PA9
+NVIC.TIM1_UP_TIM10_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:true
 RCC.TIM20Freq_Value=72000000
 ProjectManager.ComputerToolchain=false
 Mcu.Pin17=PG7
@@ -312,7 +310,6 @@ RCC.PLLM=4
 RCC.PLLN=96
 Mcu.Pin10=PC5
 PB7.Locked=true
-PA2.Mode=RMII
 PC15/OSC32_OUT.Signal=RCC_OSC32_OUT
 PA8.Mode=Activate_SOF_FS
 RCC.I2C2Freq_Value=48000000
@@ -327,6 +324,5 @@ PG7.GPIOParameters=GPIO_Label
 ProjectManager.LibraryCopy=1
 PG11.GPIOParameters=GPIO_Label
 PH0/OSC_IN.GPIOParameters=GPIO_Label
-PC1.Mode=RMII
 PA7.Signal=ETH_CRS_DV
 isbadioc=false
