+++ User/UserInc/Tasks/Task1.h
@@ -1,0 +1,13 @@
+/*
+ * Task1.h
+ *
+ *  Created on: Jun 4, 2021
+ *      Author: frank
+ */
+
+#ifndef USERINC_TASKS_TASK1_H_
+#define USERINC_TASKS_TASK1_H_
+
+void Task1_Run(void const * argument);
+
+#endif /* USERINC_TASKS_TASK1_H_ */
