+++ User/UserInc/Tasks/Task2.h
@@ -1,0 +1,13 @@
+/*
+ * Task2.h
+ *
+ *  Created on: Jun 4, 2021
+ *      Author: frank
+ */
+
+#ifndef USERINC_TASKS_TASK2_H_
+#define USERINC_TASKS_TASK2_H_
+
+void Task2_Interrupt();
+
+#endif /* USERINC_TASKS_TASK2_H_ */
