+++ User/UserInc/Tasks/Task3.h
@@ -1,0 +1,13 @@
+/*
+ * Task3.h
+ *
+ *  Created on: Jun 4, 2021
+ *      Author: frank
+ */
+
+#ifndef USERINC_TASKS_TASK3_H_
+#define USERINC_TASKS_TASK3_H_
+
+void Task3_Run(void const * argument);
+
+#endif /* USERINC_TASKS_TASK3_H_ */
